fix password migration scripts
