+++ packages/scripts/src/seed.ts
@@ -46,7 +46,7 @@ const compatibleExcelNanoid = (): string => {
   return id
 }
 
-const userPassword = '$argon2id$v=19$m=4096,t=12,p=4$P7aFyCRZ4ZOTFplVOwZGkA$bqtIRXIVrqUatyASMxPElOvJu5z5fx2bMyCK7jROIDGgHNV2MoyobQ'
+const userPassword = '$argon2id$v=19$m=4096,t=12,p=4$snysIHKmfhvtO+uORE8jwA$KemeOraiG6zasIT8bgYljyEuoGHgl5RQAUhr9bRBTVhJ11sGaErGuA'
 
 interface EventGroup {
   event: Array<CustomEvent>
